requirements.txt
dash~=1.17
dash_bootstrap_components~=0.10.3
plotly~=4.4.1
scikit-image~=0.15.0
scikit-learn~=0.21.3
sklearn~=0.0
pyarrow==1.0.0
flask~=1.1.1
Flask-Caching~=1.9.0
pandas~=1.0.5
outlier-utils~=0.0.3
python-dotenv
# imageai deps
tensorflow==2.4.0
keras==2.4.3
numpy==1.19.3
pillow==7.0.0
scipy==1.4.1
h5py==2.10.0
matplotlib==3.3.2
opencv-python~=4.2.0.34
keras-resnet==0.2.0
imageai==2.1.6
pycocotools==2.0.2
tqdm~=4.46.1
nltk~=3.5
dash_daq
alibi-detect
outlier_utils
Cython
torch
pytest
torchvision
